+++ core/src/main/scala/portals/application/task/OutputCollectorImpl.scala
@@ -18,7 +18,7 @@ private[portals] class OutputCollectorImpl[T, U, Req, Rep] extends OutputCollect
       portal: String,
       askingTask: String,
       req: Req,
-      key: Key[Int],
+      key: Key[Long],
       id: Int
   ): Unit =
     _asks = Ask(key, PortalMeta(portal, askingTask, id), req) :: _asks
@@ -33,7 +33,7 @@ private[portals] class OutputCollectorImpl[T, U, Req, Rep] extends OutputCollect
       r: Rep,
       portal: String,
       askingTask: String,
-      key: Key[Int],
+      key: Key[Long],
       id: Int
   ): Unit = _reps = Reply(key, PortalMeta(portal, askingTask, id), r) :: _reps
   def getRepOutput(): List[Reply[Rep]] = _reps.reverse
